Change SPRT ELO configuration from int to float for more precise testing

+++ src/qapla-settings.cpp
@@ -364,8 +364,8 @@ void QaplaSettings::readSprtConfig() {
     }
 
     m_sprtConfig = std::make_unique<SprtConfig>(SprtConfig{
-        .eloUpper = sprt->get<int>("eloUpper"),
-        .eloLower = sprt->get<int>("eloLower"),
+        .eloUpper = static_cast<float>(sprt->get<int>("eloUpper")),
+        .eloLower = static_cast<float>(sprt->get<int>("eloLower")),
         .alpha = sprt->get<double>("alpha"),
         .beta = sprt->get<double>("beta"),
         .maxGames = sprt->get<unsigned int>("maxgames"),
